fix: $defaultImageUrl not loading when getImageUrl() is null
ImageEntry and ImageColumn documentation states that defaultImageUrl()
is for displaying a placeholder image if one doesn't exist yet.

But If $stateItem is a filled string that points to a location that does
not exist in storage then getImageUrl() will return null due to the file not
existing and by doing so no image is loaded into the browser even though
we set a default placeholder.

So by using the ?? operator in the branch where string is filled we make
sure the browser loads an image whether it's the placeholder or the
storage one.

Sources:
https://filamentphp.com/docs/3.x/tables/columns/image#adding-a-default-image-url
https://filamentphp.com/docs/3.x/infolists/entries/image#adding-a-default-image-url